move webpack files into their own directory
