+++ webpack/dev.js
@@ -1,6 +1,6 @@
 const webpack = require('webpack');
 const merge = require('webpack-merge');
-const base = require('./webpack.base.js');
+const base = require('./base.js');
 
 module.exports = merge(base, {
   devServer: {
