+++ webpack/prod.js
@@ -1,7 +1,7 @@
 const path = require('path');
 const webpack = require('webpack');
 const merge = require('webpack-merge');
-const base = require('./webpack.base.js');
+const base = require('./base.js');
 
 module.exports = merge(base, {
   output: {
